+++ fheco/src/trs/matching_term.cpp
@@ -8,11 +8,11 @@ namespace fheco_trs
 size_t MatchingTerm::term_id = 0;
 
 MatchingTerm::MatchingTerm(int64_t _value) : term_type(TermType::scalarType), value(_value), id(term_id++) {}
-MatchingTerm::MatchingTerm(int _value) : term_type(TermType::scalarType), value(_value), id(term_id++) {}
+MatchingTerm::MatchingTerm(int _value) : term_type(TermType::scalarType), value((int64_t)_value), id(term_id++) {}
 MatchingTerm::MatchingTerm(double _value) : term_type(fheco_trs::TermType::scalarType), value(_value), id(term_id++) {}
 
 MatchingTerm::MatchingTerm(bool _value)
-  : term_type(TermType::booleanType), value(static_cast<int>(_value)), id(term_id++)
+  : term_type(TermType::booleanType), value(static_cast<int64_t>(_value)), id(term_id++)
 {}
 
 MatchingTerm::MatchingTerm(FunctionId func_id) : term_type(fheco_trs::TermType::functionType), function_id(func_id) {}
@@ -30,7 +30,6 @@ MatchingTerm::MatchingTerm(const std::string &_label, fheco_trs::TermType _term_
 
 fheco_trs::TermType MatchingTerm::deduce_term_type(fheco_trs::TermType lhs_term_type, fheco_trs::TermType rhs_term_type)
 {
-
   // Order of if statments is important !!!
 
   if (lhs_term_type == rhs_term_type)
@@ -180,7 +179,6 @@ MatchingTerm operator<<(const MatchingTerm &lhs, const MatchingTerm &rhs)
 
 MatchingTerm operator!=(const MatchingTerm &lhs, const MatchingTerm &rhs)
 {
-
   if (
     rewrite_condition_types.find(lhs.get_term_type()) == rewrite_condition_types.end() ||
     rewrite_condition_types.find(rhs.get_term_type()) == rewrite_condition_types.end())
